01_deeplab_inference-main/deeplab_inference/requirements.txt
absl-py==1.1.0
affine @ file:///Users/ktietz/demo/mc3/conda-bld/affine_1629890479180/work
anyio @ file:///C:/ci/anyio_1644463705902/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569848815/work
astunparse==1.6.3
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650292996413/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
Bottleneck @ file:///C:/ci/bottleneck_1648010921109/work
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
click-plugins @ file:///Users/ktietz/demo/mc3/conda-bld/click-plugins_1630665883852/work
cligj==0.7.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1652101770956/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///C:/ci/entrypoints_1649926659561/work
fastjsonschema @ file:///tmp/build/80754af9/python-fastjsonschema_1620414857593/work/dist
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
GDAL==3.0.2
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562631189/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
imutils==0.5.4
ipykernel @ file:///C:/ci/ipykernel_1647000985174/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1643800131373/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1644297241925/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///C:/ci/jsonschema_1650026210703/work
jupyter-client @ file:///C:/ci/jupyter_client_1650534191587/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656282279/work
jupyter-server @ file:///C:/Windows/TEMP/abs_af34539e-be0f-4767-94fb-39ff75b173480cys6cki/croots/recipe/jupyter_server_1654784029941/work
jupyterlab @ file:///opt/conda/conda-bld/jupyterlab_1647445413472/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///C:/ci/jupyterlab_server_1650271781334/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe @ file:///C:/ci/markupsafe_1654508076077/work
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
munch @ file:///Users/ktietz/demo/mc3/conda-bld/munch_1630566382504/work
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient @ file:///C:/ci/nbclient_1650308542913/work
nbconvert @ file:///C:/ci/nbconvert_1649759177374/work
nbformat @ file:///C:/ci/nbformat_1649845122517/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848126026/work
notebook @ file:///C:/ci/notebook_1651666865091/work
numexpr @ file:///C:/ci/numexpr_1640704420125/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
oauthlib==3.2.0
opencv-python==4.5.5
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1641461312795/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
progress==1.6
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyproj @ file:///C:/ci/pyproj_1614278152133/work
pyrsistent @ file:///C:/ci/pyrsistent_1636093257833/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
pywin32==302
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp37-none-win_amd64.whl
pyzmq @ file:///C:/ci/pyzmq_1638435182681/work
rasterio==1.2.10
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/ci/sniffio_1614030522573/work
snuggs @ file:///Users/ktietz/demo/mc3/conda-bld/snuggs_1629803886966/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado @ file:///C:/ci/terminado_1644322782754/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///C:/ci/urllib3_1650639842077/work
vidstab==1.7.4
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///C:/ci/websocket-client_1614804293275/work
Werkzeug==2.1.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.14.1
zipp @ file:///C:/ci/zipp_1652274072582/work
